CARDS-1484 / CARDS-1349: Move the import config into two config files for their respective uses. Add the versioning checkin to used nodes, and split the import task into two classes since it was running into fan-out complexity errors
